Internal dependencies get traced and rebuilt.
For example if an exe requires a library, then the library is an
implicit dependency to exe. Ninja manages changes and rebuilds the
library when it gets out of date.

Furthermore when a .dll is specified as an library dependency, we juggle
the name so the actual link line the compiler recieves is the .lib (since
you can't link against a .dll).