Make paths relative to project root.
This commit adds `projectRoot` field to the target schema.
This allows a developer to spectify where the project is in relation to
the target build.
This allows directories to be specified as if running from the project
root rather than from the build dir. For example it allows a src dir to
be specified as `srcDirs = ["src"]` rather than `srcDirs =
["../../srcDirs"]`.